+++ src/main/java/com/shop/concurrency/member/model/domain/Member.java
@@ -1,5 +1,6 @@
 package com.shop.concurrency.member.model.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.shop.concurrency.order.domain.Orders;
 import jakarta.persistence.*;
 import jakarta.persistence.criteria.Order;
@@ -20,10 +21,10 @@ public class Member {
     @Column(name = "member_id")
     private Long id;
 
+    @JsonIgnore
     private String name;
 
-    //1(Member) : N(Orders) -> 나는 1개고, order는 여러개다.
     @OneToMany(mappedBy = "member")// 1 : N 연관관계의 주인이 member이다.
-    List<Orders> orders = new ArrayList<>();
+    List<Orders> orders = new ArrayList<>(); //1(Member) : N(Orders) -> 나는 1개고, order는 여러개다.
 
 }
