+++ src/main/java/com/shop/concurrency/member/model/dto/response/FindMembersResponse.java
@@ -1,6 +1,9 @@
 package com.shop.concurrency.member.model.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.shop.concurrency.order.domain.Orders;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.*;
 
 @Getter
@@ -14,4 +17,5 @@ public class FindMembersResponse {
 
     private String name;
 
+    private List<Orders> orders;
 }
