+++ src/main/java/com/shop/concurrency/order/domain/Orders.java
@@ -1,5 +1,6 @@
 package com.shop.concurrency.order.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.shop.concurrency.member.model.domain.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -25,6 +26,7 @@ public class Orders {// Order는 자바 예약어이기 때문에, orders로 주
     private Long id;
 
     //N(Order) : 1(Member) -> 나(주문)는 여러개고, Member는 1개다.
+    @JsonIgnore
     @ManyToOne
     Member member;
 }
